+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `epipack.interactive`: contains a class that adds an interactive widget to Jupyter notebooks
   with which one may control the parameter values of a SymbolicEpiModel instance
+- `epipack.temporal_networks`: set up temporal networks and model simulations on them
 
 ## [v0.0.5] - 2020-08-14