+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+
+## [v0.1.0] - 2020-10-21
 
 ### Added
 - `epipack.interactive`: contains a class that adds an interactive widget to Jupyter notebooks
@@ -56,7 +58,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - initialized
 
-[Unreleased]: https://github.com/benmaier/epipack/compare/v0.0.5...HEAD
+[Unreleased]: https://github.com/benmaier/epipack/compare/v0.1.0...HEAD
+[v0.1.0]: https://github.com/benmaier/epipack/compare/v0.0.5...v0.1.0]
 [v0.0.4]: https://github.com/benmaier/epipack/compare/v0.0.4...v0.0.5]
 [v0.0.4]: https://github.com/benmaier/epipack/compare/v0.0.3...v0.0.4]
 [v0.0.3]: https://github.com/benmaier/epipack/compare/v0.0.2...v0.0.3]